chore: rm unused functions
