+++ ui/src/common/components/Search/components/QueryBuilder/utils.js
@@ -12,112 +12,6 @@ export function mergedQueries(queries) {
   return obj;
 }
 
-function query(key, value, cb, shouldOrMust = "should") {
-  if (Array.isArray(key)) {
-    return { bool: { [shouldOrMust]: key.map((k) => cb(k, value)) } };
-  }
-  return cb(key, value);
-}
-
-export const defaultOperators = [
-  {
-    value: "==",
-    text: "equals",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ term: { [k]: v } })),
-  },
-  {
-    value: "!=",
-    text: "not equals",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ bool: { must_not: { term: { [k]: v } } } }), "must"),
-  },
-  {
-    value: ">=",
-    text: "greater than or equals to",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { gte: v } } })),
-  },
-  {
-    value: "<=",
-    text: "lesser than or equals to",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { lte: v } } })),
-  },
-  {
-    value: ">",
-    text: "greater than",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { gt: v } } })),
-  },
-  {
-    value: "<",
-    text: "lesser than",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { lt: v } } })),
-  },
-  {
-    value: "∃",
-    text: "exists",
-    useInput: false,
-    query: (key) =>
-      query(key, null, (k) => ({
-        bool: {
-          // Must exists ...
-          must: { exists: { field: k } },
-          // ... and must be not empty.
-          must_not: { term: { [k]: "" } },
-        },
-      })),
-  },
-  {
-    value: "!∃",
-    text: "does not exist",
-    useInput: false,
-    query: (key) =>
-      query(
-        key,
-        null,
-        (k) => ({
-          bool: {
-            // Should be ...
-            should: [
-              // ... empty string ...
-              { term: { [k]: "" } },
-              // ... or not exists.
-              { bool: { must_not: { exists: { field: k } } } },
-            ],
-          },
-        }),
-        "must"
-      ),
-  },
-  {
-    value: "*",
-    text: "contains",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ wildcard: { [k]: `*${v}*` } })),
-  },
-  {
-    value: "!*",
-    text: "does not contains",
-    useInput: true,
-    query: (key, value) =>
-      value && query(key, value, (k, v) => ({ bool: { must_not: { wildcard: { [k]: `*${v}*` } } } }), "must"),
-  },
-  {
-    value: "^",
-    text: "start with",
-    useInput: true,
-    query: (key, value) => value && query(key, value, (k, v) => ({ wildcard: { [k]: `${v}*` } })),
-  },
-];
-
-export const defaultCombinators = [
-  { value: "AND", text: "AND" },
-  { value: "OR", text: "OR" },
-];
-
 export function uuidv4() {
   return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, (c) => {
     const r = (Math.random() * 16) | 0,
@@ -129,3 +23,229 @@ export function uuidv4() {
 export function withUniqueKey(rules) {
   return rules.map((r) => ({ ...r, key: uuidv4() }));
 }
+
+// This function transforms a text to a french compatible regex.
+// So this:
+// "Voilà un château"
+// Turns to that:
+// "[Vv][oôöOÔÖ][iïîIÏÎ]l[àâäaÀÂÄA] [uùûüUÙÛÜ]n [cçÇC]h[àâäaÀÂÄA]t[éèêëeÉÈÊËE][àâäaÀÂÄA][uùûüUÙÛÜ]"
+// It works (TM).
+function toFrenchRegex(text) {
+  return text
+    .replace(/[éèêëeÉÈÊËE]/g, "[éèêëeÉÈÊËE]")
+    .replace(/[àâäaÀÂÄA]/g, "[àâäaÀÂÄA]")
+    .replace(/[cçÇC]/g, "[cçÇC]")
+    .replace(/[iïîIÏÎ]/g, "[iïîIÏÎ]")
+    .replace(/[oôöOÔÖ]/g, "[oôöOÔÖ]")
+    .replace(/[uùûüUÙÛÜ]/g, "[uùûüUÙÛÜ]")
+    .replace(/([bdfghjklmnpqrstvwxz])/gi, (w, x) => `[${x.toUpperCase()}${x.toLowerCase()}]`);
+}
+
+function notStrict(value) {
+  return toFrenchRegex(escapeRegex(value));
+}
+
+function escapeRegex(value) {
+  return value.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&");
+}
+
+function suggestionQuery(key, value) {
+  if (Array.isArray(key)) {
+    return;
+  }
+  return {
+    query: { match_all: {} },
+    aggs: { [key]: { terms: { field: key, include: value, order: { _count: "desc" }, size: 10 } } },
+    size: 0,
+  };
+}
+
+function query(key, value, cb, shouldOrMust = "should") {
+  if (Array.isArray(key)) {
+    console.log({ bool: { [shouldOrMust]: key.map((k) => cb(k, value)) } });
+    return { bool: { [shouldOrMust]: key.map((k) => cb(k, value)) } };
+  }
+  return cb(key, value);
+}
+
+export const operators = [
+  {
+    value: "===",
+    text: "égal à (recherche stricte)",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ term: { [k]: v } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: value },
+  },
+  {
+    value: "!==",
+    text: "différent de (recherche stricte)",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ bool: { must_not: { term: { [k]: v } } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $ne: value } },
+  },
+  {
+    value: "===*",
+    text: "contient (recherche stricte)",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ wildcard: { [k]: `*${v}*` } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `.*${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`.*${escapeRegex(value)}.*`) } },
+  },
+  {
+    value: "!==*",
+    text: "ne contient pas (recherche stricte)",
+    useInput: true,
+    query: (key, value) =>
+      value && query(key, value, (k, v) => ({ bool: { must_not: { wildcard: { [k]: `*${v}*` } } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `.*${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $not: { $regex: new RegExp(`.*${escapeRegex(value)}.*`) } } },
+  },
+  {
+    value: "===^",
+    text: "commence par (recherche stricte)",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ wildcard: { [k]: `${v}*` } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`^${escapeRegex(value)}`) } },
+  },
+  {
+    value: "===$",
+    text: "termine par (recherche stricte)",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ wildcard: { [k]: `*${v}` } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `.*${escapeRegex(value)}`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`${escapeRegex(value)}$`) } },
+  },
+  {
+    value: "==",
+    text: "égal à",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ regexp: { [k]: notStrict(v) } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${notStrict(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`${notStrict(value)}`) } },
+  },
+  {
+    value: "!=",
+    text: "différent de",
+    useInput: true,
+    query: (key, value) =>
+      value && query(key, value, (k, v) => ({ bool: { must_not: { regexp: { [k]: notStrict(v) } } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${notStrict(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $not: { $regex: new RegExp(`${notStrict(value)}`) } } },
+  },
+  {
+    value: "*",
+    text: "contient",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ regexp: { [k]: `.*${notStrict(v)}.*` } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `.*${notStrict(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`.*${notStrict(value)}.*`) } },
+  },
+  {
+    value: "!*",
+    text: "ne contient pas",
+    useInput: true,
+    query: (key, value) =>
+      value &&
+      query(key, value, (k, v) => ({
+        bool: { must_not: { regexp: { [k]: `.*${notStrict(v)}.*` } } },
+      })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `.*${notStrict(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $not: { $regex: new RegExp(`.*${notStrict(value)}.*`) } } },
+  },
+  {
+    value: "^",
+    text: "commence par",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ regexp: { [k]: `${notStrict(v)}.*` } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${notStrict(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`^${notStrict(value)}`) } },
+  },
+  {
+    value: "$",
+    text: "termine par",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ regexp: { [k]: `.*${notStrict(v)}` } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `.*${notStrict(value)}`),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(`${notStrict(value)}$`) } },
+  },
+  {
+    value: ">=",
+    text: "supérieur ou égal à",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { gte: v } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $gte: value } },
+  },
+  {
+    value: "<=",
+    text: "inférieur ou égal à",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { lte: v } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $lte: value } },
+  },
+  {
+    value: ">",
+    text: "strictement supérieur à",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { gt: v } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $gt: value } },
+  },
+  {
+    value: "<",
+    text: "strictement inférieur à",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ range: { [k]: { lt: v } } })),
+    suggestionQuery: (key, value) => suggestionQuery(key, `${escapeRegex(value)}.*`),
+    mongoQuery: (key, value) => value && { [key]: { $lt: value } },
+  },
+  {
+    value: "∃",
+    text: "existe",
+    useInput: false,
+    query: (key) =>
+      query(key, null, (k) => ({
+        bool: {
+          // Must exists ...
+          must: { exists: { field: k } },
+          // ... and must be not empty.
+          must_not: { term: { [k]: "" } },
+        },
+      })),
+    mongoQuery: (key, value) => value && { [key]: { $exists: true } },
+  },
+  {
+    value: "!∃",
+    text: "n'existe pas",
+    useInput: false,
+    query: (key) =>
+      query(key, null, (k) => ({
+        bool: {
+          // Should be ...
+          should: [
+            // ... empty string ...
+            { term: { [k]: "" } },
+            // ... or not exists.
+            { bool: { must_not: { exists: { field: k } } } },
+          ],
+        },
+      })),
+    mongoQuery: (key, value) => value && { [key]: { $exists: false } },
+  },
+  {
+    value: "*.",
+    text: "regexp",
+    useInput: true,
+    query: (key, value) => value && query(key, value, (k, v) => ({ regexp: { [k]: `${v}` } })),
+    mongoQuery: (key, value) => value && { [key]: { $regex: new RegExp(value) } },
+  },
+];
+
+export const combinators = [
+  { value: "AND", text: "ET" },
+  { value: "OR", text: "OU" },
+];
